requirements.txt
tensorflow-cpu 
sklearn
urllib3==1.26.6
pandas 
numpy
gunicorn
matplotlib
streamlit
